Redo 001 easy clojure; promote to lein project
I decided to inline everything, and do away with macros. It's now just a
single 'mad-lib' function, which takes a string with some "{queries}" in
it, pulls them out, asks them of the user (removing the {}s), and
replaces the queries with the user's responses. If the user enters
nothing, it will ask again per query until an answer is given (you can
always ctrl+c to get out). The function returns the answer, so you can
do what you like with it, including printing it out. Finally, if you
pass an optional file path, the story will be saved there at the end.