Add sortDel
This gets weirder still. I can't really pop from something that's been
sorted, so I'll have to just get the head of the stack list, then delete
that value from the sort list. It is still deleting in a stack-based
way, if you stretch things a bit; it deletes the first occurrence (i.e.
the closest occurrence to the head of the list), which should be the one
added most recently, meaning this is a 'stable' system.