Add newWorld function
I don't like the idea of providing a list of blocks to this function; it
complicates creation of the world in a few ways. For one thing, it puts
the onus on the caller to create the list of blocks, when the user may
simply want an empty world, i.e. all Air. For another, it requires the
caller to provide exactly the right number of blocks, i.e. l*w*h, which
is both annoying and error-prone. I want creation of the world to be a
total function. Providing negative numbers results in an empty world,
which, from a design standpoint makes enough sense to me to skip out on
handling them explicitly. I want "design" of the space to take place via
updating of the [initially] empty world, with a map update function
quietly ignoring out-of-bounds keys.