Factor genCells out of newWorld
I had this in earlier, but it merged into newWorld before it could be
immortalized in a commit. I like the idea again, both as the type has
become simpler with the addition of the XYZ alias, but also because I
can now do things like walk every cell, which should come in handy.