Rename XYZ type to Cell; simplify World type
I like the look of Cell better than the initialism XYZ, and I like how
it starts to create Cell -> Block pairings. It's convenient in one way
to have the world size carried around by the world, but it turns out to
be very inconvenient in several other ways, and as the Data.Map.keys
function returns a sorted list of keys, and as my keys are Cells, i.e.
(x,y,z) triples, and as Int triples sort in such a way that I'll always
be able to pull the last one from the end of such a sorted list, I've
decided to simplify for expressiveness and neatness.