Add collapse1 and collapse functions
It would be more efficient to walk a core (vertical stack of blocks)
just one time, exchanging a column of Air with a column of Sand above
it, or filling it in with Lava if the the column is capped on top with
such (both of which would require keeping pointers while walking). This
could be done in a single walk. However, it would be significantly more
finicky to code up. This method has collapse1 walk the core, collapsing
things just one level; an Air-Sand pairing (Sand on top) swap places,
and an Air below a Lava becomes a Lava. This plays to the strengths of
recursive list-walking in Haskell, and makes for small, simple, readable
code. It also means that another function - collapse - must come through
and call collapse1 over and over, until everything has finally settled.
This is O(n×m), where n is the number of blocks being iterated, and m is
the number of "settlings" that need to occur amongst them. I'm certain I
can get this down to O(n), or similar, but I'll leave that for later.