Reverse direction of collapse1 function
By working from the top down (so the blocks would need to be passed in
in the opposite order now, with the first element in the list being the
top block) - far fewer collapse iterations need to be called in the case
where we have many Air blocks under a Sand or Lava block. This doesn't
preclude the need to iterate via collapse until things settle, however,
as, e.g., only the bottom-most Sand block will drop all the way down;
any above it will remain (dropping to the bottom in one pass themselves
in subsequent calls). This also happens with Lava, where Sand falling
out from below it frees it to fill the Air beneath it in a later pass.