Replace collapse functions with grouping version
This is another step closer to a single walk of the core (i.e. the list
of Blocks), but may still not be as efficient as carefully walking each
individual block and keeping pointers to things. It is able to collapse
the list without iteration, though, which is good. The idea here is to
first group like Blocks together, then walk the list of groups, moving
any Air group below any Sand group, and to convert any Air group to all
Lava, if it's followed by a Lava group. Though not obvious in the code,
the Air-under-Sand case injects the Air group back into the recursion so
it can be converted to Lava in the event that there were any Lava blocks
following the Sand block that was just moved out of the way. The thing
I'm unsure of is whether or not peeking into the front of each list on
each recursion is efficient or not.