Add genCore function
I played with the idea of this function taking the world, and figuring
out the top cell at a given xy position, but later liked the simplicity
of passing in a top Cell, and just generating the core Cells from there
to the ground. I considered naming this something else, as "core" feels
like the actual thing extracted from the ground, and this is more the
generator of the plan of where to core, but genThing keeps in line with
the earlier genCells function, which also trusts that you'll provide it
the maximum position of the world, from which to generate all cells.