Add sampleCore function
I wish there was a way in the types to know that it was impossible to
pass in a position that isn't in the world. This takes the world, and an
X and Y, and finds the topmost block at that point in the world. I don't
trust the caller to pass in the top block (I don't trust them to pass in
valid, in-world X and Y values either, but that's kind of the minimal
requirement here). Because the caller can request an out-of-bounds point
in the map, I must return the list of Blocks wrapped in a maybe, which I
do through sequencing a regular mapping. mapMaybe and catMaybes are bad
ideas here, as I wouldn't know if blocks were elided, and I wouldn't be
able to sync the core back up with the cells (which is still a puzzle).