Merge genCore function into sampleCore function
I couldn't find a good way to separate these without the possibility of
core and sample getting out of sync with each other, or [what I felt to
be] more hand-holding than was worth it. I didn't like asking the caller
to provide the world-top/highest Z value (I don't even like that I have
to let them provide the X and the Y, but they must), and you have to
pass in the world anyway, so I get the top from that. The core is
deterministic, if the X and Y are in-bounds (though there's no check),
but the sample is non-deterministic, so I map those Block lookups with a
function that returns a Maybe Block, which means I can sequence the list
of all the lookups for a Nothing if any of them are Nothing, which then
means I can - and do - fmap zipping the deterministic core with the non-
deterministic sample to get either a proper pairing, or Nothing.