Add inRadiusPoints function
This took some fighting with number types. Once you send a number into a
function, even just as a helper method, or to check on things like
bounds, the usage of what you sent in can tie that number to a type,
which can propagate back to the caller and change numbers into specific
number types that you don't want. Even my Int hints in the signature
didn't help, but at least I got a bunch of errors. I want my X and Y
values to remain integers, because I'm using them as keys in map, but
the hypot function's radius check - which uses sqrt - locks the type
down, and that propagates back to the caller, and suddenly my map keys
are Fractional (making me wonder if XY pairs as map keys isn't just a
broken idea from the start - very suspect now). The solution was to use
fromIntegral in the call to hypot, to split the types at that point.