Add circWidths function
This [hopefully efficiently] creates a list of discrete, half-widths of
the y layers of half of a 2D circle, from the center point up (or down).
It does this by walking [0..] in y, and starting at r, reducing x until
in the given radius, then incrementing y and walking from the previous x
inward. The number of distance checks is kept to a minimum in this way.