Begin Main.hs with machine and ins/push left/right
The TuringMachine data type is a first stab, as is the machine function
intended as its public constructor. I've made inserting and pushing
versions of left and right, which take the 'infinite' element to add if
the list zipper is at either end. This way I don't have to deal with an
infinite list. A lazy list, e.g. Zipper (repeat 0) (repeat 0) is nice,
but then you just end up working on the other side when trying to find
the end points for printing and such.