Add pileup function
This was one of my earlier ideas for filling up the piles; sort the
list, then just fill up the front, and keep spilling over each step of
numbers. I had another idea to move back and forth through the zipper to
take out about half of the n-complexity (restting with start over and
over), but could not get it to work. The zig-zags get in their own way,
and I couldn't find a sequence of turns and successors that built the
list in the nice way that simply resetting every time you hit a higher
number or the end of the list does.